add model file and update training

+++ training.py
@@ -13,7 +13,8 @@ from Utils.Losses import FocalLoss, dice_loss
 def save_model(model, epoch, output_path, best=False):
     if best:
         previous_best_pt = glob.glob(os.path.join(output_path, '*BEST.pt'))
-        os.remove(previous_best_pt[0])
+        if len(previous_best_pt) > 0:
+            os.remove(previous_best_pt[0])
         name = os.path.join(output_path, 'Model_stateDict__Epoch={}__BEST.pt'.format(epoch))
     else:
         name = os.path.join(output_path, 'Model_stateDict__Epoch={}.pt'.format(epoch))
@@ -27,29 +28,32 @@ def init_weights(m):
 
 
 def train(model, train_dataloader, val_dataloader, epochs, criterion_loss, optimizer, scheduler, output_path):
-    train_loss_list, val_loss_list =[], []
+    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+
+    train_loss_list, val_loss_list = [], []
     best_train_loss, best_val_loss = np.inf, np.inf
     if not os.path.isdir(os.path.join(output_path, 'saved_checkpoints')):
         os.makedirs(os.path.join(output_path, 'saved_checkpoints'))
 
     model.train()
+    model.to(device)
     softmax = nn.Softmax(dim=1)
 
     for epoch in epochs:
         running_loss = 0.0
         # --- TRAIN:  --- #
         for i, data in enumerate(train_dataloader):
-            inputs, labels = data['image'], data['gt']
+            inputs, labels = data['image'].to(device), data['gt'].to(device)
             optimizer.zero_grad()
-            outputs = model(inputs)
+            outputs = model(inputs.type(torch.FloatTensor).to(device))
             outputs = softmax(outputs)
-            loss = criterion_loss(outputs, labels)
+            loss = criterion_loss(outputs, labels.squeeze().cuda().long())
             dice_score = dice_loss(output=outputs, target=labels)  # TODO: add dice loss to total loss
             loss.backward()
             optimizer.step()
             running_loss += loss.item()
-        train_loss = running_loss / i
-        train_loss_list.append(train_loss.data.item())
+        train_loss = running_loss / (i + 1)
+        train_loss_list.append(train_loss)
         ##################
 
         model.eval()
@@ -57,26 +61,26 @@ def train(model, train_dataloader, val_dataloader, epochs, criterion_loss, optim
         # --- VAL:  --- #
         val_running_loss = 0.0
         for val_i, val_data in enumerate(val_dataloader):
-            inputs, labels = val_data['image'], val_data['gt']
-            outputs = model(inputs)
+            inputs, labels = val_data['image'].to(device), val_data['gt'].to(device)
+            outputs = model(inputs.type(torch.FloatTensor).to(device))
             outputs = softmax(outputs)
-            loss = criterion_loss(outputs, labels)
+            loss = criterion_loss(outputs, labels.squeeze().cuda().long())
             dice_score = dice_loss(output=outputs, target=labels)  # TODO: add dice loss to total loss
             val_running_loss += loss.item()
-        val_loss = val_running_loss / val_i
-        val_loss_list.append(val_loss.data.item())
+        val_loss = val_running_loss / (val_i + 1)
+        val_loss_list.append(val_loss)
         ##################
 
         # --- Save results to csv ---#
         current_lr = optimizer.param_groups[-1]['lr']
         current_wd = optimizer.param_groups[-1]['wd']
-        epoch_results = [epoch, train_loss.data.item(), val_loss.data.item(), current_lr, current_wd]
+        epoch_results = [epoch, train_loss, val_loss, current_lr, current_wd]
         write_to_csv(os.path.join(output_path, 'results.csv'), epoch_results)
         ##############################
 
         # --- Save model checkpoint ---#
-        if val_loss.item() < best_val_loss:
-            best_val_loss = val_loss.item()
+        if val_loss < best_val_loss:
+            best_val_loss = val_loss
             save_model(model=model, epoch=epoch, output_path=os.path.join(output_path, 'saved_checkpoints'), best=True)
         elif epoch % 10:
             save_model(model=model, epoch=epoch, output_path=os.path.join(output_path, 'saved_checkpoints'), best=False)
@@ -94,22 +98,23 @@ def main():
     config.read('config.ini')
 
     batch_size = config.getint('Params', 'batch_size')
-    lr = config.getfloat('Params','lr')
+    lr = config.getfloat('Params', 'lr')
     wd = config.getfloat('Params', 'wd')
-    epochs= config.getint('Params', 'epochs')
+    epochs = config.getint('Params', 'epochs')
     output_folder = config['Paths']['output_folder']
 
     # --- Dataset --- #
-    datasets_imgs_folders = {'train': config['input_images']['train'],
-                             'val': config['input_images']['val'],
-                             'test': config['input_images']['test']}
-    gt_imgs_folders = {'train': config['ground_truth']['train'],
-                             'val': config['ground_truth']['val'],
-                             'test': config['ground_truth']['test']}
+    train_images = glob.glob(os.path.join(config['input_images']['train'], '*/*.jpg'))
+    val_images = glob.glob(os.path.join(config['input_images']['val'], '*/*.jpg'))
+    train_gt = glob.glob(os.path.join(config['ground_truth']['train'], '*/*.jpg'))
+    val_gt = glob.glob(os.path.join(config['ground_truth']['val'], '*/*.jpg'))
 
-    train_dataloader, val_dataloader, test_dataloader = get_dataloaders(dataset_folders=datasets_imgs_folders,
-                                                                        gt_folders=gt_imgs_folders,
-                                                                        batch_size=batch_size, num_workers=4)
+    datasets_imgs_folders = {'train': train_images, 'val': val_gt}
+    gt_imgs_folders = {'train': val_images, 'val': train_gt}
+
+    train_dataloader, val_dataloader, _ = get_dataloaders(dataset_dict=datasets_imgs_folders,
+                                                          gt_dict=gt_imgs_folders,
+                                                          batch_size=batch_size, num_workers=4)
     #####################
 
     # --- Model --- #
@@ -123,7 +128,7 @@ def main():
     #####################
 
     # --- Loss function --- #
-    loss_BCE = nn.BCELoss()
+    loss_bce = nn.BCELoss()
     focal_loss = FocalLoss(gamma=2, alpha=1)
     optimizer = torch.optim.Adam(filter(lambda p: p.requires_grad, model.parameters()), lr=lr, weight_decay=wd)
     scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, 'min', patience=10)
